+++ CHANGELOG.md
@@ -38,6 +38,8 @@
 
 ### Fixed
 
+- [#4959](https://github.com/ChainSafe/forest/pull/4959) Re-enable garbage
+  collection after implementing a "persistent" storage for manifests.
 - [#4988](https://github.com/ChainSafe/forest/pull/4988) Fix the `logs` member
   in `EthTxReceipt` that was initialized with a default value.